Creation of checks to prevent calling Adyen API multiple times If the Adyen Express module is enabled then they will have already gathered the payment methods to be added into the customer data.
Added a check to see if any of the configuration for Adyen Express is enabled and if so we will not run the lookup ourselves but instead use a plugin to take the data that already exists and use it for our purposes.
