+++ etc/di.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0"?>
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
     <type name="Magento\Checkout\CustomerData\Cart">
-        <plugin name="addAdyenMethodsToCartData" type="BlueFinch\CheckoutAdyen\Plugin\CustomerData\Cart" />
+        <plugin name="addAdyenMethodsToCartData" type="BlueFinch\CheckoutAdyen\Plugin\CustomerData\AddAdyenExpressMethods" />
+    </type>
+    <type name="Adyen\ExpressCheckout\Plugin\CustomerData\Cart">
+        <plugin name="BlueFinch\CheckoutAdyen\Plugin\CustomerData\Cart" type="BlueFinch\CheckoutAdyen\Plugin\CustomerData\Cart" />
     </type>
 </config>
